# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: EventShooter API - Premium event photography booking platform
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /photographers:
    get:
      operationId: listPhotographers
      tags: [photographers]
      summary: List photographers with filters
      parameters:
        - name: city
          in: query
          schema: { type: string }
        - name: eventType
          in: query
          schema: { type: string }
        - name: minPrice
          in: query
          schema: { type: number }
        - name: maxPrice
          in: query
          schema: { type: number }
        - name: rating
          in: query
          schema: { type: number }
        - name: search
          in: query
          schema: { type: string }
        - name: page
          in: query
          schema: { type: integer }
        - name: limit
          in: query
          schema: { type: integer }
      responses:
        "200":
          description: Photographers list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PhotographerListResponse"
...
  /photographers/{id}:
    get:
      operationId: getPhotographer
      tags: [photographers]
      summary: Get photographer profile
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Photographer
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PhotographerProfile"
    patch:
      operationId: updatePhotographer
      tags: [photographers]
      summary: Update photographer profile
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PhotographerUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PhotographerProfile"
  /photographers/{id}/portfolio:
    get:
      operationId: getPortfolio
      tags: [photographers]
      summary: Get photographer portfolio items
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Portfolio
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PortfolioItem"
    post:
      operationId: addPortfolioItem
      tags: [photographers]
      summary: Add portfolio item
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PortfolioItemInput"
      responses:
        "201":
          description: Added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PortfolioItem"
  /photographers/{id}/availability:
    get:
      operationId: getPhotographerAvailability
      tags: [photographers]
      summary: Get photographer availability calendar
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Availability
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AvailabilitySlot"
    post:
      operationId: setPhotographerAvailability
      tags: [photographers]
      summary: Set photographer availability
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AvailabilityInput"
      responses:
        "200":
          description: Set
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /photographers/{id}/packages:
    get:
      operationId: getPhotographerPackages
      tags: [photographers]
      summary: Get photographer packages/pricing
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Packages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Package"
    post:
      operationId: createPackage
      tags: [photographers]
      summary: Create a new package
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PackageInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Package"
  /photographers/{id}/earnings:
    get:
      operationId: getPhotographerEarnings
      tags: [photographers]
      summary: Get photographer earnings summary
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Earnings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EarningsSummary"
  /photographers/{id}/dashboard:
    get:
      operationId: getPhotographerDashboard
      tags: [photographers]
      summary: Photographer dashboard overview
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Dashboard
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PhotographerDashboard"

  # ─── BOOKINGS ───
...
  /payments/initiate:
    post:
      operationId: initiatePayment
      tags: [payments]
      summary: Initiate Razorpay/UPI payment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PaymentInitInput"
      responses:
        "200":
          description: Payment initiated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaymentInitResponse"
  /payments/verify:
    post:
      operationId: verifyPayment
      tags: [payments]
      summary: Verify payment after Razorpay callback
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PaymentVerifyInput"
      responses:
        "200":
          description: Verified
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Payment"
  /payments/{id}/invoice:
    get:
      operationId: getInvoice
      tags: [payments]
      summary: Get invoice for a payment
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Invoice
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Invoice"
  /payments/{id}/refund:
    post:
      operationId: requestRefund
      tags: [payments]
      summary: Request refund for a payment
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RefundInput"
      responses:
        "200":
          description: Refund requested
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"

  # ─── REVIEWS ───
...
components:
  schemas:
...
    MessageResponse:
      type: object
      required: [message]
      properties:
        message: { type: string }

    # ─── AUTH SCHEMAS ───
...
    User:
      type: object
      required: [id, name, email, role, createdAt]
      properties:
        id: { type: integer }
        name: { type: string }
        email: { type: string }
        phone: { type: ["string", "null"] }
        avatarUrl: { type: ["string", "null"] }
        role: { type: string, enum: [customer, photographer, admin] }
        isVerified: { type: boolean }
        isBanned: { type: boolean }
        city: { type: ["string", "null"] }
        referralCode: { type: ["string", "null"] }
        createdAt: { type: string }
...
    Photographer:
      type: object
      required: [id, userId, displayName, city, rating, totalReviews, startingPrice, isAvailable, isFeatured]
      properties:
        id: { type: integer }
        userId: { type: integer }
        displayName: { type: string }
        bio: { type: ["string", "null"] }
        city: { type: string }
        specializations: { type: array, items: { type: string } }
        rating: { type: number }
        totalReviews: { type: integer }
        totalBookings: { type: integer }
        startingPrice: { type: number }
        coverImageUrl: { type: ["string", "null"] }
        avatarUrl: { type: ["string", "null"] }
        isAvailable: { type: boolean }
        isFeatured: { type: boolean }
        isVerified: { type: boolean }
        isApproved: { type: boolean }
        yearsOfExperience: { type: ["integer", "null"] }
    PhotographerProfile:
      type: object
      required: [id, userId, displayName, city, rating, totalReviews, startingPrice, isAvailable, isFeatured]
      properties:
        id: { type: integer }
        userId: { type: integer }
        displayName: { type: string }
        bio: { type: ["string", "null"] }
        city: { type: string }
        specializations: { type: array, items: { type: string } }
        rating: { type: number }
        totalReviews: { type: integer }
        totalBookings: { type: integer }
        startingPrice: { type: number }
        coverImageUrl: { type: ["string", "null"] }
        avatarUrl: { type: ["string", "null"] }
        isAvailable: { type: boolean }
        isFeatured: { type: boolean }
        isVerified: { type: boolean }
        isApproved: { type: boolean }
        yearsOfExperience: { type: ["integer", "null"] }
        portfolio: { type: array, items: { $ref: "#/components/schemas/PortfolioItem" } }
        packages: { type: array, items: { $ref: "#/components/schemas/Package" } }
        reviews: { type: array, items: { $ref: "#/components/schemas/Review" } }
        equipment: { type: ["string", "null"] }
        instagramHandle: { type: ["string", "null"] }
    PhotographerListResponse:
      type: object
      required: [photographers, total, page, limit]
      properties:
        photographers:
          type: array
          items:
            $ref: "#/components/schemas/Photographer"
        total: { type: integer }
        page: { type: integer }
        limit: { type: integer }
    PhotographerUpdate:
      type: object
      properties:
        displayName: { type: string }
        bio: { type: string }
        city: { type: string }
        specializations: { type: array, items: { type: string } }
        startingPrice: { type: number }
        coverImageUrl: { type: string }
        avatarUrl: { type: string }
        isAvailable: { type: boolean }
        yearsOfExperience: { type: integer }
        equipment: { type: string }
        instagramHandle: { type: string }
    PortfolioItem:
      type: object
      required: [id, photographerId, mediaUrl, mediaType, createdAt]
      properties:
        id: { type: integer }
        photographerId: { type: integer }
        mediaUrl: { type: string }
        mediaType: { type: string, enum: [photo, video, reel] }
        title: { type: ["string", "null"] }
        eventType: { type: ["string", "null"] }
        isApproved: { type: boolean }
        createdAt: { type: string }
    PortfolioItemInput:
      type: object
      required: [mediaUrl, mediaType]
      properties:
        mediaUrl: { type: string }
        mediaType: { type: string, enum: [photo, video, reel] }
        title: { type: string }
        eventType: { type: string }
    AvailabilitySlot:
      type: object
      required: [id, photographerId, date, isAvailable]
      properties:
        id: { type: integer }
        photographerId: { type: integer }
        date: { type: string }
        isAvailable: { type: boolean }
        timeSlots: { type: array, items: { type: string } }
    AvailabilityInput:
      type: object
      required: [dates]
      properties:
        dates: { type: array, items: { type: string } }
        isAvailable: { type: boolean }
    Package:
      type: object
      required: [id, photographerId, name, price, duration]
      properties:
        id: { type: integer }
        photographerId: { type: integer }
        name: { type: string }
        description: { type: ["string", "null"] }
        price: { type: number }
        duration: { type: integer }
        includes: { type: array, items: { type: string } }
        addons: { type: array, items: { $ref: "#/components/schemas/PackageAddon" } }
        eventTypes: { type: array, items: { type: string } }
        isActive: { type: boolean }
    PackageAddon:
      type: object
      required: [id, name, price]
      properties:
        id: { type: integer }
        name: { type: string }
        price: { type: number }
        description: { type: ["string", "null"] }
    PackageInput:
      type: object
      required: [name, price, duration]
      properties:
        name: { type: string }
        description: { type: string }
        price: { type: number }
        duration: { type: integer }
        includes: { type: array, items: { type: string } }
        eventTypes: { type: array, items: { type: string } }
    EarningsSummary:
      type: object
      required: [totalEarned, pendingAmount, thisMonth, lastMonth]
      properties:
        totalEarned: { type: number }
        pendingAmount: { type: number }
        thisMonth: { type: number }
        lastMonth: { type: number }
        monthlyData: { type: array, items: { $ref: "#/components/schemas/MonthlyEarning" } }
    MonthlyEarning:
      type: object
      required: [month, amount, bookings]
      properties:
        month: { type: string }
        amount: { type: number }
        bookings: { type: integer }
    PhotographerDashboard:
      type: object
      required: [totalBookings, pendingBookings, completedBookings, totalEarnings, rating, totalReviews]
      properties:
        totalBookings: { type: integer }
        pendingBookings: { type: integer }
        confirmedBookings: { type: integer }
        completedBookings: { type: integer }
        totalEarnings: { type: number }
        thisMonthEarnings: { type: number }
        rating: { type: number }
        totalReviews: { type: integer }
        recentBookings: { type: array, items: { $ref: "#/components/schemas/Booking" } }
...
    Booking:
      type: object
      required: [id, customerId, photographerId, eventType, eventDate, city, totalAmount, advanceAmount, status, createdAt]
      properties:
        id: { type: integer }
        customerId: { type: integer }
        photographerId: { type: integer }
        packageId: { type: ["integer", "null"] }
        eventType: { type: string }
        eventDate: { type: string }
        eventEndDate: { type: ["string", "null"] }
        city: { type: string }
        venue: { type: ["string", "null"] }
        totalAmount: { type: number }
        advanceAmount: { type: number }
        remainingAmount: { type: number }
        status: { type: string, enum: [pending, confirmed, rejected, completed, cancelled, in_progress] }
        notes: { type: ["string", "null"] }
        referenceImages: { type: array, items: { type: string } }
        selectedAddons: { type: array, items: { type: string } }
        rejectionReason: { type: ["string", "null"] }
        cancellationReason: { type: ["string", "null"] }
        photographer: { $ref: "#/components/schemas/Photographer" }
        customer: { $ref: "#/components/schemas/User" }
        package: { $ref: "#/components/schemas/Package" }
        createdAt: { type: string }
        updatedAt: { type: string }
...
    Payment:
      type: object
      required: [id, bookingId, amount, type, status, createdAt]
      properties:
        id: { type: integer }
        bookingId: { type: integer }
        amount: { type: number }
        type: { type: string, enum: [advance, remaining, full, refund] }
        status: { type: string, enum: [pending, completed, failed, refunded] }
        method: { type: ["string", "null"] }
        razorpayOrderId: { type: ["string", "null"] }
        razorpayPaymentId: { type: ["string", "null"] }
        createdAt: { type: string }
    PaymentInitInput:
      type: object
      required: [bookingId, amount, type]
      properties:
        bookingId: { type: integer }
        amount: { type: number }
        type: { type: string, enum: [advance, remaining, full] }
        method: { type: string, enum: [razorpay, upi] }
    PaymentInitResponse:
      type: object
      required: [orderId, amount, currency]
      properties:
        orderId: { type: string }
        amount: { type: number }
        currency: { type: string }
        keyId: { type: string }
    PaymentVerifyInput:
      type: object
      required: [bookingId, razorpayOrderId, razorpayPaymentId, razorpaySignature]
      properties:
        bookingId: { type: integer }
        razorpayOrderId: { type: string }
        razorpayPaymentId: { type: string }
        razorpaySignature: { type: string }
    Invoice:
      type: object
      required: [id, bookingId, paymentId, invoiceNumber, amount, issuedAt]
      properties:
        id: { type: integer }
        bookingId: { type: integer }
        paymentId: { type: integer }
        invoiceNumber: { type: string }
        amount: { type: number }
        taxAmount: { type: number }
        customerName: { type: string }
        photographerName: { type: string }
        eventType: { type: string }
        eventDate: { type: string }
        issuedAt: { type: string }
    RefundInput:
      type: object
      required: [reason]
      properties:
        reason: { type: string }
        amount: { type: number }

    # ─── REVIEW SCHEMAS ───
    Review:
      type: object
      required: [id, customerId, photographerId, bookingId, rating, createdAt]
      properties:
        id: { type: integer }
        customerId: { type: integer }
        photographerId: { type: integer }
        bookingId: { type: integer }
        rating: { type: number }
        comment: { type: ["string", "null"] }
        isApproved: { type: boolean }
        customer: { $ref: "#/components/schemas/User" }
        createdAt: { type: string }
...